shackle: Replace this package with direct use of `display-buffer-alist'
I thought I wanted this package but it was the root cause of some
problems I've been seeing over the last few years.

The biggest problem I encountered was with the `(:same t)` setting
allowing buffers to take over the minibuffer.  Once that happens the
current frame became unusable.

I tried for a few hours to fix this and then gave up and wrote the new
`buffers.el` file as a replacement.