build: Switch to a Makefile and Nix Flakes
With the new Makefile, we can build in parallel, and it's *so* much
faster!

Also switched to Nix Flakes.  I'm hoping this makes it easier to use
this package in my home configuration and make it easier to do
upgrades.

I still need to convert the remaining niv sources into flake inputs.