consult: Prevent `consult-buffer' and friends from starting new tramp connections
If `default-directory' is a remote file name at the time
`consult-buffer' is started, tramp will try to start a new connection.
This mostly happens to me in vterm when I've used ssh to get to a
remote host, and then want to switch buffers.

Maybe related to https://github.com/minad/consult/issues/224

Note: this issue with tramp doesn't happen with the built in
`switch-to-buffer' function.