org: Remove org-superstar-mode due to bugs
This mode seems to interact with the org cache in some weird way and
triggers a bunch of warnings when a file changes behind org's back.

I also disabled this mode for a few days and decided I actually like
the plain text heading and item markers.  They also have more
consistent scaling on HiDPI monitors.