aggressive-indent: Purge this package from my config
It's quite annoying how this package destroys the indentation below
point while you are in the middle of writing an expression.  It does
end up putting the indentation back though.  But in some odd cases I
want to indent things in a specific way (or a formatting tool does)
and this mode fights against that.